packages/web_crawler/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "wealth_management_portal_web_crawler" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/packages/web_crawler/html'
 --cov-report xml:'../../coverage/packages/web_crawler/coverage.xml'
 --html='../../reports/packages/web_crawler/unittests/html/index.html'
 --junitxml='../../reports/packages/web_crawler/unittests/junit.xml'
"""

[tool.hatch.build.targets.wheel]
packages = [ "wealth_management_portal_web_crawler" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[tool.uv.sources.wealth-management-portal-common-market-events]
workspace = true
editable = true

[tool.uv.sources."wealth_management_portal.common_auth"]
workspace = true

[project]
name = "wealth_management_portal.web_crawler"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "wealth_management_portal.common_auth",
  "feedparser>=6.0.12",
  "beautifulsoup4>=4.14.3",
  "requests>=2.32.5",
  "newspaper3k>=0.2.8",
  "python-dateutil>=2.9.0.post0",
  "wealth-management-portal-common-market-events",
  "lxml-html-clean>=0.4.3",
  "mcp==1.26.0",
  "strands-agents",
  "boto3==1.42.44",
  "aws-opentelemetry-distro==0.15.0"
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"
